finished(?) Connect Four.py
There are bound to be some bugs. If you encounter an Error message, or
the game doesn’t recognize that a player has 4 in a row (or any other
bug), please copy and paste everything in the output window into a
textedit file, and send me that file! Also, I am open to suggestions
for improvement.