added Connect Four tinter.py
I’m attempting to make a GUI for the Connect Four game via the Tkinter
module. It’s still a work in progress, and right now I am just going to
try and make feature by feature compatible. This is my first time using
Tkinter, so it might take a while. Right now, you can place a piece on
the board by clicking a button (thats a big improvement!). Next I will
add a function to check for a win.